@@ -5,19 +5,19 @@
   "dependencies": {
     "@material-ui/core": "^4.11.0",
     "@material-ui/icons": "^4.9.1",
-    "@nivo/scatterplot": "^0.62.0",
-    "axios": "^0.19.2",
+    "@nivo/scatterplot": "^0.63.0",
+    "axios": "^0.20.0",
     "file-saver": "^2.0.2",
-    "jwt-decode": "^2.2.0",
-    "material-ui-popup-state": "^1.6.1",
-    "react": "^16.13.1",
+    "jwt-decode": "^3.0.0",
+    "material-ui-popup-state": "^1.7.0",
+    "react": "^17.0.0",
     "react-avatar": "^3.9.7",
-    "react-dom": "^16.13.1",
-    "react-notifications": "^1.6.0",
+    "react-dom": "^17.0.0",
+    "react-notifications": "^1.7.2",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "^3.4.3",
-    "react-scroll-section": "^3.0.0",
-    "timeago-react": "^3.0.0"
+    "react-scripts": "^3.4.4",
+    "react-scroll-section": "^3.0.1",
+    "timeago-react": "^3.0.1"
   },
   "scripts": {
     "start": "react-scripts start",
