@@ -48,34 +48,37 @@ def create_app(config_class=Config):
         os.mkdir(app.config['TMP_FOLDER'])
 
     if not app.debug and not app.testing:
-        if app.config['MAIL_SERVER']:
-            auth = None
-            if app.config['MAIL_USERNAME'] or app.config['MAIL_PASSWORD']:
-                auth = (app.config['MAIL_USERNAME'],
-                        app.config['MAIL_PASSWORD'])
-            secure = None
-            if app.config['MAIL_USE_TLS']:
-                secure = ()
-            mail_handler = SMTPHandler(
-                mailhost=(app.config['MAIL_SERVER'], app.config['MAIL_PORT']),
-                fromaddr='no-reply@' + app.config['MAIL_SERVER'],
-                toaddrs=app.config['ADMINS'], subject='oka Failure',
-                credentials=auth, secure=secure)
-            mail_handler.setLevel(logging.ERROR)
-            app.logger.addHandler(mail_handler)
+        app.logger = logging.getLogger("oka_logger")
+        if not app.logger.handlers:
+            if app.config['MAIL_SERVER']:
+                auth = None
+                if app.config['MAIL_USERNAME'] or app.config['MAIL_PASSWORD']:
+                    auth = (app.config['MAIL_USERNAME'],
+                            app.config['MAIL_PASSWORD'])
+                secure = None
+                if app.config['MAIL_USE_TLS']:
+                    secure = ()
+                mail_handler = SMTPHandler(
+                    mailhost=(app.config['MAIL_SERVER'],
+                              app.config['MAIL_PORT']),
+                    fromaddr='no-reply@' + app.config['MAIL_SERVER'],
+                    toaddrs=app.config['ADMINS'], subject='oka Failure',
+                    credentials=auth, secure=secure)
+                mail_handler.setLevel(logging.ERROR)
+                app.logger.addHandler(mail_handler)
 
-        if not os.path.exists('logs'):
-            os.mkdir('logs')
-        file_handler = RotatingFileHandler('logs/oka.log',
-                                           maxBytes=10240, backupCount=10)
-        file_handler.setFormatter(logging.Formatter(
-            '%(asctime)s %(levelname)s: %(message)s '
-            '[in %(pathname)s:%(lineno)d]'))
-        file_handler.setLevel(logging.INFO)
-        app.logger.addHandler(file_handler)
+            if not os.path.exists('logs'):
+                os.mkdir('logs')
+            file_handler = RotatingFileHandler('logs/oka.log',
+                                               maxBytes=10240, backupCount=10)
+            file_handler.setFormatter(logging.Formatter(
+                '%(asctime)s %(levelname)s: %(message)s '
+                '[in %(pathname)s:%(lineno)d]'))
+            file_handler.setLevel(logging.INFO)
+            app.logger.addHandler(file_handler)
 
-        app.logger.setLevel(logging.INFO)
-        app.logger.info('oka startup')
+            app.logger.setLevel(logging.INFO)
+            app.logger.info('oka startup')
 
     return app
 
